Rework views
Added films container and film item views, edited film list view
accordingly.

Added no films view.
