+++ src/view/films-container-view.js
@@ -1,0 +1,23 @@
+import { createElement } from 'utils';
+
+const createFilmsContainerTemplate = () => '<div class="films-list__container"></div>';
+
+export default class FilmsContainerView {
+  #element = null;
+
+  get template() {
+    return createFilmsContainerTemplate();
+  }
+
+  get element() {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
+    }
+
+    return this.#element;
+  }
+
+  removeElement() {
+    this.#element = null;
+  }
+}
