+++ src/view/index.js
@@ -1,6 +1,9 @@
 export { default as FilmListView } from './film-list-view.js';
+export { default as FilmsContainerView } from './films-container-view.js';
+export { default as FilmItemView } from './film-item-view.js';
 export { default as FilterView } from './filter-view.js';
 export { default as FilmsCountView } from './films-count-view.js';
+export { default as NoFilmsView } from './no-films-view.js';
 export { default as MainContentView } from './main-content-view.js';
 export { default as ShowMoreButtonView } from './show-more-button-view.js';
 export { default as SortView } from './sort-view.js';
